% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{analyze.simstudy.lm}
\alias{analyze.simstudy.lm}
\title{analyze.simstudy.lm}
\usage{
analyze.simstudy.lm(
  the.coefs,
  summary.stats,
  conf.level = 0.95,
  the.quantiles = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975),
  coef.name = "Coefficient",
  estimate.name = "Estimate",
  lm.p.name = "Pr(>|t|)",
  f.p.name = "f.pvalue"
)
}
\arguments{
\item{the.coefs}{A data.frame or data.table object of the summary table of estimated coefficients for repeated linear regression models.  Structure is in the form returned by the function simulation::sim.statistics.lm$the.coefs().}

\item{summary.stats}{A data.frame or data.table object of the summary statistics of repeated linear regression models.  Structure is in the form returned by the function simitation::sim.statistics.lm$summary.stats().}

\item{conf.level}{A numeric value between 0 and 1 representing the confidence level (1 - significance level).}

\item{the.quantiles}{A numeric vector of values between 0 and 1.  Summary statistics to analyze the tests will return the specified quantiles.}

\item{coef.name}{A character value specifying the column of the.coefs that contains the names of the input variables of the linear regression model.}

\item{estimate.name}{A character value specifying the column of the.coefs that contains the estimated coefficients of the linear regression model.}

\item{lm.p.name}{A character value specifying the column of the.coefs that contains the p-values for the tests of the estimated coefficients of the linear regression model.}

\item{f.p.name}{A character value specifying the column of summary.stats that contains the p-value for the linear regression model's F test.}
}
\value{
Returns a list with four elements:
1. 'lm.estimate.summary': A data.table that provides a summary of the estimates of the coefficients for each variable in the linear model. The summary includes several descriptive statistics such as the mean, standard error, and various quantiles as specified by the 'the.quantiles' argument.
2. 'lm.p.summary': A data.table that shows the proportion of times the null hypothesis was rejected (i.e., p-value is less than 1 - 'conf.level') and not rejected (i.e., p-value is greater than or equal to 1 - 'conf.level') for each variable in the model.
3. 'lm.stats.summary': A data.table that provides a summary of several overall statistics of the linear model, including sigma (standard error of the residuals), rse (residual standard error), r.squared, adjusted r.squared, and F statistic. The summary includes the mean, standard error, and the specified quantiles for each of these statistics.
4. 'fstatistic.p.summary': A data.table that provides the proportion of times the null hypothesis was rejected (i.e., p-value is less than 1 - 'conf.level') and not rejected (i.e., p-value is greater than or equal to 1 - 'conf.level') for the overall F statistic of the model.
}
\description{
analyze.simstudy.lm
}
