% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{analyze.simstudy.logistic}
\alias{analyze.simstudy.logistic}
\title{analyze.simstudy.logistic}
\usage{
analyze.simstudy.logistic(
  the.coefs,
  summary.stats,
  conf.level = 0.95,
  the.quantiles = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975),
  coef.name = "Coefficient",
  estimate.name = "Estimate",
  logistic.p.name = "Pr(>|z|)"
)
}
\arguments{
\item{the.coefs}{A data.frame or data.table object of the summary table of estimated coefficients for repeated logistic regression models.  Structure is in the form returned by the function simitation::sim.statistics.logistic$the.coefs().}

\item{summary.stats}{A data.frame or data.table object of the summary statistics of repeated logistic regression models.  Structure is in the form returned by the function simitation::sim.statistics.logistic$summary.stats().}

\item{conf.level}{A numeric value between 0 and 1 representing the confidence level (1 - significance level).}

\item{the.quantiles}{A numeric vector of values between 0 and 1.  Summary statistics to analyze the tests will return the specified quantiles.}

\item{coef.name}{A character value specifying the column of the.coefs that contains the names of the input variables of the logistic regression model.}

\item{estimate.name}{A character value specifying the column of the.coefs that contains the estimated coefficients of the logistic regression model.}

\item{logistic.p.name}{A character value specifying the column of the.coefs that contains the p-values for the tests of the estimated coefficients of the logistic regression model.}
}
\value{
Returns a list with three elements:
1. 'logistic.estimate.summary': A data.table that provides a summary of the estimate of coefficients for each variable in the logistic regression model. The summary includes several descriptive statistics such as the mean, standard error, and various quantiles as specified by the 'the.quantiles' argument.
2. 'logistic.p.summary': A data.table that shows the proportion of times the null hypothesis was rejected (i.e., p-value is less than 1 - 'conf.level') and not rejected (i.e., p-value is greater than or equal to 1 - 'conf.level') for each variable in the model.
3. 'logistic.stats.summary': A data.table that provides a summary of several overall statistics of the logistic regression model such as deviance, Akaike Information Criterion (AIC), degrees of freedom for the residuals, null deviance, degrees of freedom for the null model, number of iterations, and dispersion. The summary includes the mean, standard error, and the specified quantiles for each of these statistics.
}
\description{
analyze.simstudy.logistic
}
