% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{buildsim.binary}
\alias{buildsim.binary}
\title{buildsim.binary}
\usage{
buildsim.binary(the.formula, the.variable, n, num.experiments = 1)
}
\arguments{
\item{the.formula}{A character string specifying the probability of success in the form of a function (e.g., "binary(0.5)" for a 0.5 probability of success).}

\item{the.variable}{A character string specifying the name of the variable to be used in the output data frame.}

\item{n}{The number of rows (samples) to be generated in the output data frame.}

\item{num.experiments}{The number of experiments to be conducted, resulting in multiple sets of binary data. The default is 1, indicating a single set of binary data.}
}
\value{
The function returns a data frame containing binary values, where each row represents a binary outcome based on the specified probability of success.
}
\description{
This internal function generates binary data based on a specified probability of success. The probability of success is extracted from the formula provided as a character string (e.g., binary(0.5) indicates a 0.5 probability of success). The function then returns a data frame containing binary values with the specified number of rows (n) and repetitions (num.experiments).
}
