% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{buildsim.binomial}
\alias{buildsim.binomial}
\title{buildsim.binomial}
\usage{
buildsim.binomial(the.formula, the.variable, n, num.experiments = 1)
}
\arguments{
\item{the.formula}{A character string specifying the binomial distribution in the form "Bin(num.trials, success.prob)", where num.trials is the number of trials for each observation, and success.prob is the probability of success in each trial.}

\item{the.variable}{A character string specifying the name of the variable to be used in the output data frame.}

\item{n}{The number of observations to be generated in the output data frame. Each observation consists of num.experiments sets of binomial trials.}

\item{num.experiments}{The number of repetitions or experiments to be conducted, resulting in multiple sets of binomial data. The default is 1, indicating a single set of binomial data.}
}
\value{
The function returns a data frame containing the generated binomial data, where each row represents an observation with the specified number of successes out of num.trials trials.
}
\description{
This internal function generates data based on a binomial distribution. It simulates a specified number of binomial trials (success/failure experiments) with a given probability of success for each trial. The function returns a data frame containing the generated binomial data.
}
