% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{buildsim.lm}
\alias{buildsim.lm}
\title{buildsim.lm}
\usage{
buildsim.lm(dat, the.formula, the.variable, n, num.experiments = 1)
}
\arguments{
\item{dat}{The input data for the simulation, which can be either a data.table or data.frame.}

\item{the.formula}{A character string specifying the linear regression formula in the form of "lm(fixed.terms + residual.error)", where fixed.terms represents the fixed components of the formula, and residual.error specifies the residual error term.}

\item{the.variable}{A character string specifying the name of the variable to be used in the output data frame.}

\item{n}{The number of observations to be generated in the output data frame. Each observation consists of num.experiments sets of simulated data.}

\item{num.experiments}{The number of repetitions or experiments to be conducted, resulting in multiple sets of simulated data. The default is 1, indicating a single set of simulated data.}
}
\value{
The function returns a data frame containing the generated data based on the specified linear regression formula. Each row in the data frame represents an observation with the simulated values based on the formula. The column name of the generated data is specified by the the.variable parameter.
}
\description{
This internal function generates simulated data for linear regression models. It uses a linear regression formula with fixed terms and a specification of the residual error. The function then returns a data frame containing the generated data based on the specified formula.
}
