% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{buildsim.logistic}
\alias{buildsim.logistic}
\title{buildsim.logistic}
\usage{
buildsim.logistic(dat, the.formula, the.variable, n, num.experiments = 1)
}
\arguments{
\item{dat}{The input data for the simulation, which can be either a data.table or data.frame.}

\item{the.formula}{A character string specifying the logistic regression formula in the form of "logistic(fixed.terms)", where fixed.terms represents the fixed components of the formula.}

\item{the.variable}{A character string specifying the name of the variable to be used in the output data frame.}

\item{n}{The number of observations to be generated in the output data frame. Each observation consists of num.experiments sets of simulated data.}

\item{num.experiments}{The number of repetitions or experiments to be conducted, resulting in multiple sets of simulated data. The default is 1, indicating a single set of simulated data.}
}
\value{
The function returns a data frame containing the generated binary outcomes (success or failure) based on the specified logistic regression formula. Each row in the data frame represents an observation with the simulated binary outcome. The column name of the generated data is specified by the the.variable parameter.
}
\description{
This internal function generates simulated data for logistic regression models. It uses a logistic regression formula with fixed terms to compute the log-odds, and then applies the logistic function to obtain the probabilities of success (binary outcomes). The function returns a data frame containing the generated binary outcomes based on the specified logistic regression formula.
}
