% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{buildsim.normal}
\alias{buildsim.normal}
\title{buildsim.normal}
\usage{
buildsim.normal(the.formula, the.variable, n, num.experiments)
}
\arguments{
\item{the.formula}{A character string specifying the normal distribution formula in the form of "N(mean, sd)", where "mean" represents the mean of the normal distribution, and "sd" represents the standard deviation.}

\item{the.variable}{A character string specifying the name of the variable to be used in the output data frame.}

\item{n}{The number of observations to be generated in the output data frame. Each observation consists of num.experiments sets of simulated data.}

\item{num.experiments}{The number of repetitions or experiments to be conducted, resulting in multiple sets of simulated data.}
}
\value{
The function returns a data frame containing the generated data based on the specified normal distribution formula. Each row in the data frame represents an observation with the simulated values from the normal distribution. The column name of the generated data is specified by the the.variable parameter.
}
\description{
This internal function generates simulated data from a normal distribution. It takes a normal distribution formula in the form of "N(mean, sd)" and returns a data frame containing the simulated data based on the specified mean and standard deviation.
}
