% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{buildsim.poisson}
\alias{buildsim.poisson}
\title{buildsim.poisson}
\usage{
buildsim.poisson(the.formula, the.variable, n, num.experiments)
}
\arguments{
\item{the.formula}{A character string specifying the Poisson distribution formula in the form of "poisson(lambda)", where "lambda" represents the mean parameter of the Poisson distribution. The function is sensitive to the specific format and will raise errors if it is not complied with.}

\item{the.variable}{A character string specifying the name of the column to be used in the output data frame.}

\item{n}{The number of samples to be generated for each experiment.}

\item{num.experiments}{The number of independent sets of 'n' samples to be generated, effectively multiplying the total number of observations generated by 'n'.}
}
\value{
The function returns a data frame containing the generated Poisson-distributed data as specified by the input parameters. The data frame has 'n * num.experiments' rows and a single column named as per the.variable parameter, each row representing a single observation.
}
\description{
This internal function generates simulated data from a Poisson distribution. It expects a formula specification in the form of "poisson(lambda)" and returns a data frame containing simulated data based on the specified mean parameter lambda.
}
