% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{buildsim.sample}
\alias{buildsim.sample}
\title{buildsim.sample}
\usage{
buildsim.sample(
  the.formula,
  the.variable,
  n,
  num.experiments,
  value.split = ",",
  symbol.open.paren = "(",
  symbol.close.paren = ")"
)
}
\arguments{
\item{the.formula}{A character string specifying the sample distribution formula, which contains the possible values in the first set of parentheses and their respective probabilities in the second set. The function will raise an error if the formula is not correctly formatted.}

\item{the.variable}{A character string specifying the name of the column to be used in the output data frame.}

\item{n}{The number of samples to be generated for each experiment.}

\item{num.experiments}{The number of independent sets of 'n' samples to be generated, effectively multiplying the total number of observations generated by 'n'.}

\item{value.split}{A character string specifying the delimiter used to separate the values in the formula. The default is a comma.}

\item{symbol.open.paren}{A character string specifying the character used to open the parentheses in the formula. The default is an open parenthesis.}

\item{symbol.close.paren}{A character string specifying the character used to close the parentheses in the formula. The default is a close parenthesis.}
}
\value{
The function returns a data frame containing the generated sample-distributed data as specified by the input parameters. The data frame has 'n * num.experiments' rows and a single column named as per the.variable parameter, each row representing a single observation.
}
\description{
This internal function generates simulated data from a specified sample distribution. It expects a formula specification in the form "sample(('Red', 'Green', 'Blue'), (0.5, 0.3, 0.2))" where the first parentheses set corresponds to possible values, and the second corresponds to their respective probabilities. The function returns a data frame containing simulated data.
}
