% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{buildsim.uniform}
\alias{buildsim.uniform}
\title{buildsim.uniform}
\usage{
buildsim.uniform(the.formula, the.variable, n, num.experiments = 1)
}
\arguments{
\item{the.formula}{A character string specifying the uniform distribution formula in the form of "U(min,max)", where "min" and "max" represent the minimum and maximum parameters of the uniform distribution. The function will raise an error if the formula is not correctly formatted.}

\item{the.variable}{A character string specifying the name of the column to be used in the output data frame.}

\item{n}{The number of samples to be generated for each experiment.}

\item{num.experiments}{The number of independent sets of 'n' samples to be generated, effectively multiplying the total number of observations generated by 'n'. Default is 1.}
}
\value{
The function returns a data frame containing the generated uniformly-distributed data as specified by the input parameters. The data frame has 'n * num.experiments' rows and a single column named as per the.variable parameter, each row representing a single observation.
}
\description{
This internal function generates simulated data from a uniform distribution. It expects a formula specification in the form of "U(min,max)" and returns a data frame containing simulated data.
}
