% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{identify.distribution}
\alias{identify.distribution}
\title{identify.distribution}
\usage{
\method{identify}{distribution}(
  dat = NULL,
  the.step,
  n,
  num.experiments,
  step.split = "~",
  value.split = ","
)
}
\arguments{
\item{dat}{A data frame or data.table object. Default is NULL. When provided, it is combined with the output data frame generated by the simulation.}

\item{the.step}{A character string specifying the formula and the variable in the format "variable ~ distribution".}

\item{n}{The number of samples to be generated for each experiment.}

\item{num.experiments}{The number of independent sets of 'n' samples to be generated, effectively multiplying the total number of observations generated by 'n'.}

\item{step.split}{A character string specifying the delimiter used to separate the variable and the distribution in the.step. Default is "~".}

\item{value.split}{A character string specifying the delimiter used to separate the values in the formula. The default is a comma.}
}
\value{
The function returns a data frame or data.table (if dat was provided) containing the simulated data as specified by the input parameters. The column name of the simulated data is the variable provided in the.step.
}
\description{
This internal function identifies the type of distribution specified in the provided formula, and then calls the appropriate function to generate simulated data based on this distribution. The function supports normal, uniform, poisson, binary, binomial, sample (categorical), logistic, and linear regression (lm) distributions. It expects a formula specification in the form of "dist(param1, param2)".
}
