% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{internal.quantiles.mean.sd}
\alias{internal.quantiles.mean.sd}
\title{internal.quantiles.mean.sd}
\usage{
internal.quantiles.mean.sd(x, the.quantiles, na.rm = T)
}
\arguments{
\item{x}{A numeric vector for which quantiles, the mean, and standard deviation will be computed.}

\item{the.quantiles}{A numeric vector of probabilities with values in [0,1] representing the quantiles to be computed. If not provided or empty, the default values are c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975).}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds. Default is TRUE.}
}
\value{
The function returns a data.table containing the computed quantiles (named as 'q.[quantile]'), the mean ('mean'), and the standard deviation ('st.error') of the numeric vector x.
}
\description{
This internal function computes quantiles, the mean, and the standard deviation of a given numeric vector, x. If no specific quantiles are specified, it computes the 2.5th, 10th, 25th, 50th (median), 75th, 90th, and 97.5th percentiles by default.
}
