% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{internal.statistics.one.lm}
\alias{internal.statistics.one.lm}
\title{internal.statistics.one.lm}
\usage{
internal.statistics.one.lm(the.data, the.formula)
}
\arguments{
\item{the.data}{A data.table or data.frame containing the data on which the linear regression is performed.}

\item{the.formula}{A character string specifying the formula of the linear regression model. This formula will be passed to stats::lm() for the regression analysis.}
}
\value{
A list containing two data.tables: coef.table containing the coefficients of the model and summary.stats containing various fit statistics of the model.
}
\description{
This internal utility function performs a linear regression analysis on the input data using the provided formula (the.formula). It generates a summary of the model fit, which includes coefficient estimates, sigma, degrees of freedom, residual standard error (RSE), R-squared, adjusted R-squared, F-statistic and its associated degrees of freedom and p-value. These statistics are organized into two tables and returned as a list.
}
