% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{internal.statistics.one.logistic}
\alias{internal.statistics.one.logistic}
\title{internal.statistics.one.logistic}
\usage{
internal.statistics.one.logistic(the.data, the.formula)
}
\arguments{
\item{the.data}{A data.table or data.frame containing the data on which the logistic regression is performed.}

\item{the.formula}{The formula representing the logistic regression model.}
}
\value{
A list containing two data.tables: coef.table containing the coefficients of the model and summary.stats containing various fit statistics of the model.
}
\description{
This internal utility function performs a logistic regression analysis on the input data using the provided formula (the.formula) and returns various statistics related to the regression model.
}
