% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{quantile.dt}
\alias{quantile.dt}
\title{quantile.dt}
\usage{
\method{quantile}{dt}(x, probs, na.rm = T)
}
\arguments{
\item{x}{A numeric vector whose quantiles are to be calculated.}

\item{probs}{A numeric vector of probabilities at which the quantiles will be calculated.}

\item{na.rm}{A logical value indicating whether NA values should be removed before the calculation.}
}
\value{
A data.table where the column names are the specified probabilities (prefixed with 'q.') and the values are the corresponding quantiles of x.
}
\description{
This function calculates the quantiles of a numeric vector at the specified probabilities and returns the results as a data.table.
}
