% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim.statistics.lm}
\alias{sim.statistics.lm}
\title{sim.statistics.lm}
\usage{
sim.statistics.lm(simdat, the.formula, grouping.variables)
}
\arguments{
\item{simdat}{A data.table or data.frame that holds the data to be used in multivariable
regression models across one or more experiments. The structure is as returned by the function
simitation::simulation.steps().}

\item{the.formula}{A formula object or character string specifying the formula for the
regression model. The formula represents the linear regression model to be fitted to the data.}

\item{grouping.variables}{A character vector of column names from 'simdat' on which to
group the data. The linear regression model will be fit separately for each group defined by
these variables.}
}
\value{
A list containing the regression coefficients (the.coefs) and summary statistics
(summary.stats) for each group of data.
}
\description{
This function fits a linear regression model to data grouped by given variables and
returns the regression coefficients and summary statistics for each group.
}
